requirements-test.txt
pytest==6.0.1
vcrpy==4.1.0
Flask-Testing==0.8.0